Added support for services and service groups.
There is a small issue with service groups, a minimum of 1 tag is required. It looks like this is a mistake in the provider. Until I can submit a PR, I put a bogus tag in to make it work. I have to add this tag when using this module so it exists. Once I submit a PR and the provider is fixed, this bogus tag setup will be removed.